refactoring Addons - now from Addon import * will import all addon decorators.
