+++ bots/replyphoto.py
@@ -1,6 +1,7 @@
-from helper.Addon import is_reply
+from Addon import *
 from helper.ImageHelper import ImageHelper as ih
 from irispy2 import ChatContext
+from Addon import *
 
 def reply_photo(chat: ChatContext):
     match chat.message.command:
