+++ helper/BanControl.py
@@ -1,6 +1,4 @@
-from helper.Admin import is_admin
-from helper.Addon import is_reply
-from helper.DatabaseHelper import get_reply_user_id
+from Addon import *
 from irispy2 import ChatContext
 from helper.BotManager import BotManager
