+++ helper/DatabaseHelper.py
@@ -1,7 +1,5 @@
-from irispy2 import ChatContext, Bot
 from irispy2.bot.models import Message
 from helper.BotManager import BotManager
-import json
 
 def get_reply_user_id(message: Message):
     reply_chat = get_reply_chat(message)
